closure: goog.ui.TableSorter: store header element in case columns move
If columns are dragged around in the table (for instance as a result
of the use of dragtable script), the columen index in the TableSorter
object becomes invalid.  Because of that sorting by one column, then
swapping it's place with another one, and clicking on it again causes
the order of rows to stay the same instead of being reversed.

By saving header element instead of column index in the TableSorter
object, this issue is resolved.

It is important to note that dragging columns around will still not
work with setSortFunction as it uses column indexes.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=51958852