There are times you render a textarea into an non-document attached div. When this happens, you won't get a padding box from goog.style.paddingbox. Because of this, each call to getHeight_ needs to ensure that the paddingbox exists, and if it does, create it. Once it is created, the result can be cached for all other uses.
The issue this looks like is that if the textarea is rendered before the
div is put into the document, resizing just does not work because the padding
box.top effectively returns zero, which makes the heightAfterNudge == currentHeight
always return true, which ends up with us trying to set a bogus height (like.... 3).

Not sure how to write a test for this.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=56689785