Change goog.isFoo & goog.dom.isFoo functions to take an UNKNOWN instead of an ALL typed argument.
While this will often not make a difference, it does change how type inference works when these functions are passed to i.e. the templatized version of goog.array.filter because the type of the passed in function goes into type union with the passed in array. Because the type union of anything with * is * this essentially throws away all type information.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=57709512