Support using goog.dom.classlist on non-html elements in IE 10.
The specific change is to check for the classList attribute on an individual
element instead of checking just for general classList support. This
prevents an exception seen on IE 10, which (contrary to spec) supports
classList on HTMLElement but not on SVGElement.

These calls are still not guaranteed to succeed on non-html elements:
for instance, if the element lacks a classList attribute, the fallback
className attribute may not work on SVG elements, wherein the className
is not a string. But now it's at least safe to call these functions
without risking throwing an exception, and their behavior on non-html
elements is now no worse than the behavior of goog.dom.classes.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=58125677