Fix a bug in goog.ui.Select where if an item has already been (programmatically) selected before the Select is displayed, and no default caption has been set, the selection is changed to the first item when the Select is displayed.
This is an alternate fix for the goog.ui.Select bug (first fix attempted in CL/58057704) which hopefully won't break either BizBuilder or Oz.

Note: If this CL causes a test failure or unexpected behaviour for you, it is because in your code you have called setSelectedItem before the Select is displayed, *and* you are expecting the first item to be selected when it is displayed (instead of the item you've previously selected). In that case, either the test should be updated to reflect the correct behaviour (of expecting the selected item to be displayed), or the code needs to be fixed not to select that item.

Currently all tests pass for:
i18n.address, commerce.bizbuilder, people.oz, people.plusone, closure, caribou.bigtop

http://test/OCL%3A58091250%3ABASE%3A58092632%3A1386851359736%3Ad1804e88
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=58328519