Fix: date formatting specifies parameters as goog.date.DateLike, but fails for goog.date.Date.
One could argue that is is the mistake of the caller to take try and format a goog.date.Date as date-time. But DateLike kind of hides that.
One also argue that because goog.date.Date does not have getHours and Co it really is not date-like.
But changing the DateLike would probably break too may things, so...
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=60022358