Have MockClock throw immediately on receiving a non-function, rather than failing internally later when .tick() tries to call it.
A plausible and arguably more correct alternative would be to implement the actual browser behavior, which is to eval() anything that's not a function. I figured we don't want to encourage the implicit use of eval and this was simpler.

Note: This change could cause an existing user of MockClock to fail in the edge case where it performs a bad setTimeout and then never calls .tick() afterward.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=67528677