Create goog.html.SafeStyle.create and goog.html.SafeStyle.concat.
We currently allow unconstrained "style" attribute in SafeHtml. With these functions, I'll be able to constrain it to SafeStyle.

So if we can agree that this is safe then it could be a reasonable improvement over the current state where SafeStyle is just a goog.string.Const wrapper.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=67553749