Disallow ; and } in SafeStyle const value.
I'm not adding a check for ", /* and ( which should be treated the same way in fromConstant as well as in const value until we decide whether to allow them or not. I think we should definitely allow ( and perhaps also " and it doesn't make much sense to disallow just /* as it has the same context-switching effect as the other two.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=67699385