Sets the spell checker as the parent component for the spell check suggestion menu.
If the parent is not set then focus shifts outside of the current element until user browses the suggestion menu and upon hiding it refocuses the editor which triggers a large set of announcements (especially if within a dialog). If the parent is set focus does remain within the descendants of the editor field thus reducing the amount of redundant announcements.

Tested locally with Win7 + FF 30 + NVDA and Chrome + ChromeVox
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=71239687